Completed conversion from float to BigDecimal
As described in the developer journal, a number of small but annoying
errors were found in the math for floats that, in a game of stock
trading that is heavily balanced on math and numbers, would eventually
come out to be a pain in the ass. The decision was to move to BigDecimal
numbers. This commit reflects the completion of that conversion process
for the Stock, Trader, SimConstants, and related classes.